backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-keycloak==0.24.0
python-dotenv==0.19.0
PyJWT==2.1.0
requests==2.25.1
